Proposed approach to S3 profiling refactoring
